updated target selection mechanism
safety check on the closest enemy method was removed as it became obsolete by, per default assigning the character itself to be the target.
aside from the healer, which needs to decide whther he wants to target an ally or enemy, with this minor change, assigning a default target in the unity inspector is no longer required as well, which brings a lot of convenience.